+++ config/knip.config.ts
@@ -677,6 +677,8 @@ const config = {
       // Copied as executable runtime internals by the package artifact manifest.
       "src/runtime-internals/mcp-command-line.mjs!",
       "src/runtime-internals/mcp-proxy.mjs!",
+      // Spawned by the real-process elicitation regression through CODEX_PATH.
+      "test/fixtures/codex-app-server.mjs!",
     ]),
     [`${BUNDLED_PLUGIN_ROOT_DIR}/azure-speech`]: bundledPluginWorkspace(),
     [`${BUNDLED_PLUGIN_ROOT_DIR}/browser`]: bundledPluginWorkspace([
