+++ src/agents/harness/structured-input-boundary.ts
@@ -1,5 +1,45 @@
 import { truncateUtf16Safe } from "../../utils.js";
-import type { StructuredInputRecord, StructuredInputValue } from "./structured-input.js";
+import type { AgentHarnessUserInputQuestion } from "./user-input-bridge.js";
+
+type StructuredInputScalar = string | number | boolean | null;
+export type StructuredInputValue =
+  | StructuredInputScalar
+  | StructuredInputValue[]
+  | StructuredInputRecord;
+export type StructuredInputRecord = { [key: string]: StructuredInputValue };
+export type StructuredInputAnswerValue = string | number | boolean | string[];
+
+type StructuredInputDecodeResult =
+  | { kind: "absent" }
+  | { kind: "invalid"; message: string }
+  | { kind: "present"; entries: Array<[string, StructuredInputAnswerValue]> };
+
+export type StructuredInputField = {
+  question: AgentHarnessUserInputQuestion;
+  decode: (values: readonly string[]) => StructuredInputDecodeResult;
+};
+
+type StructuredInputPlan =
+  | { kind: "form"; intro: string; fields: StructuredInputField[] }
+  | { kind: "url"; question: AgentHarnessUserInputQuestion };
+
+export type StructuredInputCompileResult =
+  | { kind: "ready"; plan: StructuredInputPlan }
+  | { kind: "unsupported"; message: string };
+
+export type StructuredInputCompilerOptions = {
+  protocolName: string;
+  allowEmptyForm?: boolean;
+  minimumChoiceCount?: 1 | 2;
+  allowEnumNames?: boolean;
+  allowImagePicker?: boolean;
+  booleanLabels?: readonly [string, string];
+  metadata?: {
+    secretPath?: readonly string[];
+    otherAnswerPath?: readonly string[];
+    otherQuestionIdPath?: readonly string[];
+  };
+};
 
 const MAX_SNAPSHOT_DEPTH = 8;
 const MAX_SNAPSHOT_NODES = 256;
